Move IIssuesFilter, its implementation and test into connectedmode assembly.
