+++ src/Integration.Vsix/AsmRef_Integration.Vsix_Baseline_WithoutStrongNames.txt
@@ -1,7 +1,7 @@
 ---
 ################################
 # Assembly references report
-# Report date/time: 2023-03-14T09:19:34.9248702Z
+# Report date/time: 2023-03-14T12:41:35.4337459Z
 ################################
 #
 # Generated by Devtility CheckAsmRefs v0.11.0.223
@@ -109,6 +109,7 @@ Referenced assemblies:
 - 'PresentationCore, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35'
 - 'PresentationFramework, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35'
 - 'SonarLint.VisualStudio.CFamily, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
+- 'SonarLint.VisualStudio.ConnectedMode, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
 - 'SonarLint.VisualStudio.Core, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
 - 'SonarLint.VisualStudio.Infrastructure.VS, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
 - 'SonarLint.VisualStudio.Integration, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
@@ -121,7 +122,7 @@ Referenced assemblies:
 - 'System.IO.Abstractions, Version=9.0.0.0, Culture=neutral, PublicKeyToken=96bf224d23c43e59'
 - 'System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
 - 'System.Xaml, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
-# Number of references: 26
+# Number of references: 27
 
 ---
 Assembly: 'SonarLint.VisualStudio.CFamily, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
@@ -266,7 +267,6 @@ Referenced assemblies:
 - 'PresentationCore, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35'
 - 'PresentationFramework, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35'
 - 'SonarLint.VisualStudio.CloudSecrets, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
-- 'SonarLint.VisualStudio.ConnectedMode, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
 - 'SonarLint.VisualStudio.Core, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
 - 'SonarLint.VisualStudio.Infrastructure.VS, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
 - 'SonarLint.VisualStudio.Progress, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
@@ -281,7 +281,7 @@ Referenced assemblies:
 - 'System.Xml, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
 - 'System.Xml.Linq, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
 - 'WindowsBase, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35'
-# Number of references: 32
+# Number of references: 31
 
 ---
 Assembly: 'SonarLint.VisualStudio.Integration.TeamExplorer, Version=6.14.0.0, Culture=neutral, PublicKeyToken=null'
